+++ bash-scripts/k8s-clusters/destroy-kops-bucket.sh
@@ -3,15 +3,26 @@
 set -e
 
 
-# setup
-source ${ROOT_PATH}/bash-scripts/devops-functions.sh
-run_setup
+if [ -d "${SOURCE_PATH}/kops-bucket" ]; then
 
-BUCKET=${KOPS_BUCKET_NAME}
-delete_versioned_bucket_contents
+    echo "${SOURCE_PATH}/kops-bucket found. Deleting..."
 
-cd ${SOURCE_PATH}/kops-bucket
-terraform init
+    # setup
+    source ${ROOT_PATH}/bash-scripts/devops-functions.sh
+    run_setup
 
-echo "Destroying kops state bucket..."
-terraform destroy --auto-approve
+    BUCKET=${KOPS_BUCKET_NAME}
+    delete_versioned_bucket_contents
+
+    cd ${SOURCE_PATH}/kops-bucket
+    terraform init
+
+    echo "Destroying kops state bucket..."
+    terraform destroy --auto-approve
+
+    echo "Removing kops-bucket files..."
+    rm -rf ${SOURCE_PATH}/kops-bucket
+
+else
+    echo "KOPS state bucket path '${SOURCE_PATH}/kops-bucket' does not exist. So let's not delete it. ;)"
+fi
