+++ bash-scripts/k8s-clusters/destroy-remote-state-resources.sh
@@ -2,16 +2,25 @@
 
 set -e
 
+if [ -d "${SOURCE_PATH}/remote-state" ]; then
 
-# setup
-source ${ROOT_PATH}/bash-scripts/devops-functions.sh
-run_setup
+    echo "${SOURCE_PATH}/remote-state found. Deleting..."
 
-BUCKET=${TERRAFORM_BUCKET_NAME}
-delete_versioned_bucket_contents
+    # setup
+    source ${ROOT_PATH}/bash-scripts/devops-functions.sh
+    run_setup
 
-cd ${SOURCE_PATH}/remote-state
-terraform init
+    BUCKET=${TERRAFORM_BUCKET_NAME}
+    delete_versioned_bucket_contents
 
-echo "Destroying remote state resources..."
-terraform destroy --auto-approve
+    cd ${SOURCE_PATH}/remote-state
+    terraform init
+
+    echo "Destroying terraform remote-state resources..."
+    terraform destroy --auto-approve
+
+    echo "Destroying remote-state files..."
+    rm -rf ${SOURCE_PATH}/remote-state
+else
+    echo "Remote-state resources path '${SOURCE_PATH}/remote-state' does not exist. So let's not delete it. ;)"
+fi
